1.新增跨平台的RfClock和RfSleep函数； 2.添加RfClock和RfSleep相关单元测试； 3.并将项目中使用clock和Sleep/usleep的地方替换为RfClock和RfSleep； 4.添加RfService和RfRunnableService的单元测试。

+++ trunk/Framework/RfService.h
@@ -52,8 +52,8 @@ protected:
 	RfServiceManager* getManager() const;
 
 private:
-	long mInitTime;
-	long mTickTime;
+	unsigned int mInitTime;
+	unsigned int mTickTime;
 	bool mIsInitialized;
 	RfServiceManager* mManager;
 };
